==> 07/PART 1 - Naive_Bayes_on_Amazon_Food_Reviews.ipynb <==
import pandas as pd
reviews = pd.read_csv('reviews.csv')

reviews.sample()

reviews.shape

# Data Preprocessing

## 1. Filter out some (Score, Summary, Text) features from dataset

df = reviews.filter(['Score','Summary','Text'], axis = 1)
df.iloc[:5]

## 2.   Splitting the data into train and test

# finding the number of rows in our dataset
rows = df.shape[0]
rows

train_size = int(rows * 0.70)
test_size = rows - train_size

rows == train_size + test_size

train_data = pd.DataFrame(columns=df.columns)
test_data  = pd.DataFrame(columns=df.columns)

print("Train Data : {}".format(train_data))
print("\nTest Data: {}".format(test_data))
print("\ntest_size : {} train_size : {}".format(test_size, train_size))

# Didn't use this function to split the data, it will take lot of time say more than 20 minutes...

"""
test_index = 0
train_index = 0

# split the data into training and testing data
for index, row in df.iterrows():
    print(index)
    if train_size is not 0 and test_size is not 0:
        if random.random()<=0.5:
            train_data.loc[train_index] = row
            train_size = train_size - 1
            train_index= train_index + 1
        else:
            test_data.loc[test_index] = row
            test_size = test_size - 1
            test_index = test_index + 1
    elif test_size is 0:
        # test_size is full. lets add this row to training data...
        train_data.loc[train_index] = row
        train_size = train_size - 1
        train_index= train_index + 1
    else:
        # train data is full. lets add that row to test data..
        test_data.loc[test_index] = row
        test_size = test_size - 1
        test_index = test_index + 1
"""

import random

# indices list contains all indices from 0 to rows..
indices =list(range(rows))
random.shuffle(indices)

# get random indices from shuffled indices array for train and test data..
trnind = indices[:train_size]
tstind = indices[train_size:rows]

print(trnind[:10])
print(tstind[:10])

# get the training data with that trnindices array
train_data = df.iloc[trnind]
train_data.shape

# get the test data with that tstindices array
test_data = df.iloc[tstind]
test_data.shape

train_data.iloc[:3]

test_data.iloc[:3]

## 3. Some Preprocessing

* ___Removing html tags___

* ___Removing all special characters___

* ___Convert all text into small characters___

* ___Removing stop words___

* ___Stemming words using Porter Stemming algorithm___

import re
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer

ps = PorterStemmer()

def preProcessString(text):
    # remove all html tags
    text = re.sub('<.*?>', ' ', str(text))
    
    # remove all special characters
    text = re.sub('[^A-Za-z0-9]+', ' ', text)
    
    # converting all text into small letters and store them as words for furthur processing
    text_list = text.lower().split()
    
    # removing stopwords from the text
    english_stop_words = set(stopwords.words('english'))
    # we have used set instead of list because, set uses hashing to store the words. So lookup is O(1).
    # where as for list the look up time is O(n) (ie., make things faster in list comprehension below)
    text_list = [word for word in text_list if word not in english_stop_words]
    
    # stemming the words (removing prefix and postfix) using Porter stemming algorithm..
    text_list = [ps.stem(word) for word in text_list]
    
    return ' '.join(text_list)

train_data['Summary'] = train_data['Summary'].apply(preProcessString)

train_data['Text'] = train_data['Text'].apply(preProcessString)

test_data['Summary'] = test_data['Summary'].apply(preProcessString)

test_data['Text'] = test_data['Text'].apply(preProcessString)

# let's check some random datapoint of test data
test_data.sample()

# let's check some random datapoint of train data
train_data.sample()

# combine these two datasets so that we can save them as a csv file locally..
edited_reviews = pd.concat([train_data, test_data])

edited_reviews.to_csv('preprocessed_reviews_SST.csv', index=False)

# It is same size as original dataframe. (NO DATA LOSS)
edited_reviews.shape
