==> 07/Part 2 -  Naive_Bayes_on_Preprocessed_data.ipynb <==
import pandas as pd
import numpy as np

### This csv file contains Score, Summary and Text as preprocessed data

reviews_sst = pd.read_csv('preprocessed_reviews_SST.csv')

reviews_sst[:3]

### split the data into train and test

rows = reviews_sst.shape[0]
rows

train_size = int(rows*0.70)
test_size = rows - train_size

print(train_size)
print(test_size)

# creating new dataframes for train aand test data
train_data = pd.DataFrame(columns=reviews_sst.columns)
test_data  = pd.DataFrame(columns=reviews_sst.columns)

print("Train Data : {}".format(train_data))
print("\nTest Data: {}".format(test_data))
print("\ntest_size : {} train_size : {}".format(test_size, train_size))

import random

# indices list contains all indices from 0 to rows..
indices =list(range(rows))
random.shuffle(indices)

# get random indices from shuffled indices array for train and test data..
trnind = indices[:train_size]
tstind = indices[train_size:rows]

print(trnind[:10])
print(tstind[:10])

# split the data into train and test based on the indices that we saperated
train_data = reviews_sst.iloc[trnind]
test_data = reviews_sst.iloc[tstind]

train_data[:1]

test_data[:1]

# Counter is subclass of dict
from collections import Counter
# we will update this counter whenever needed
counter = Counter()

counter.clear()

# get summary and text from train data
summary = train_data['Summary']
print(summary.shape)
text = train_data['Text']
print(text.shape)

# get word count from summaries of all train data and update to the counter
for line in summary.iteritems():
    counter.update(str(line[1]).split())

# get word count from text data of train data and update the counter
for line in text.iteritems():
    counter.update(str(line[1]).split())

len(counter)

import sys
sys.getsizeof(counter)

counter['best']

counter.most_common(2)

# reviews_sst.shape[0]
rows

score_counts = [sum(reviews_sst['Score'] == 1), \
                sum(reviews_sst['Score'] == 2), \
                sum(reviews_sst['Score'] == 3), \
                sum(reviews_sst['Score'] == 4), \
                sum(reviews_sst['Score'] == 5)]
scores = dict(zip(list('12345'), score_counts))
print(sum(score_counts) == rows)
print(scores)

# how many reviews are there that has rating
scores['1'], scores['2'], scores['3'], scores['4'], scores['5']

word_count = dict()

for tple in train_data.itertuples():
    score = str(tple[1])
    summary = str(tple[2])
    text = str(tple[3])
    words = set(text.split())
    words.update(summary.split())
    for word in words:
        if word in word_count:
            word_count[word].update(score)
        else:
            word_count[word] = Counter(score)

# the following code should be used if you want to maintain word frequecy instead of word occurance in a document.

# for word in summary.split():
#     if word in word_count:
#         word_count[word].update(score)
#     else:
#         word_count[word] = Counter(score)
# for word in text.split():
#     if word in word_count:
#         word_count[word].update(score)
#     else:
#         word_count[word] = Counter(score)
        

len(word_count)

#   word_count represents how many reviews(of specific score) that a perticular word occured. 
# It did't consider the case that the word occurred more than once in a single review. 


print(word_count['stick'])
print(word_count['great'])

# Testing Naive Bayes (Bernoulli model)  with our test data

#sample_test = test_data[:100]

test_data.head()

# Things that we will do to test our model

 1. We will consider scores (1,2) as NEGATIVE and (4,5) as POSITIVE and we will ignore reviews that has score 3.
 2. ie., we will consider scores(1,2,4,5) while building Confusion matrix.

___TP : (true positives)___ : # of reviews that are predicted POSITIVE and it is TRUE.
 
        ie., # of POSITIVE reviews that are predicted correctly (as POSITIVE)
___FP : (false positives)___ : # of reviews that are predicted as POSITIVE and it is FALSE.__(error)__
    
         ie., # of NEGATIVE reviews that are predicted incorrectly ( as POSITIVE)
___FN : (flase negatives)___ : # of reviews that are predicted as NEGATIVE and it is FALSE.__(error)__
 
         ie., # of POSITIVE reviews that are predicted incorrectly ( as NEGATIVE)
___TN : (true negative)___ : # of revies that are predicted as NEGATIVE and it is TRUE.

        ie., # of NEGATIVE reviews that are predicted correctly(as NEGATIVE)

### PRECISION

> ___  Out of all those reviews that you are predicted as +ve, HOW PRECISE ARE YOU ..?___

+ #of reviews that our model predicted as +ve = reviews that are (correctly predicted as +Ve) + ( incorrectly predicted as  +Ve)
+ #of times you correctly predicted as +Ve = TPs

$precision = \frac{TPs} {TPs+FPs}$


### RECALL

> ___Out of all those +Ve reviews, how many times you recalled(predicted) them correcly (positive).___ 

+ #of reviews that are +ve = +Ve reviews that are (correctly predicted as +Ve) + ( incorrectly predicted as  -Ve)
+ #of times you correctly predicted as +Ve = TPs

$recall = \frac{TPs} {TPs + FNs}$ 

# function to get probability
def prob(word, score):
    return (word_count[word][score] if word in word_count else 0 + 1)/(scores[score] + 2)
prob('like','4')

> posterior $ \alpha$ (prior $\times$ likelihood)

+ but, since our data is imbalanced (80% are positive reviews), we will make prior as constant(0.5). ie., we are basically ignore that prior term.

> posterior $ \alpha $ likelihood

def likelihood(words, score):
    likelihood = 1
    for word in words:
        likelihood = likelihood * prob(word, score)
    return likelihood

def update_confusion_matrix(predicted, actual):
#     print(predicted, actual)
    # change this score into positive and negative
    predicted = 'positive' if 4<=predicted<=5 else 'negative'
    actual = 'positive' if 4<=actual<=5 else 'negative'

#     print(predicted, actual)
    if predicted is actual and actual is 'positive':
        # update true positives
        confusion_matrix.update({'TP':1})
    elif predicted is actual and actual is 'negative':
        # update true negatives
        confusion_matrix.update({'TN':1})
    elif predicted is not actual and actual is 'positive':
        # update false negatives
        confusion_matrix.update({'FN':1})
    elif predicted is not actual and actual is 'negative':
        # update false positives
        confusion_matrix.update({'FP':1})

# intialize confusion matrix
confusion_matrix = Counter()

for row in test_data.itertuples():
    score, summery, text = row[1:]
    # just getting all words at one place ( in a set )
    words = set(str(text).split())
    words.update(str(summery).split())
    # get all likelihoods of a review having some score into this list
    probs = list()
    for rating in list('12345'):
        probs.append(likelihood(words, rating))
    # store our predicted score into pred_score variable
    pred_score = np.argmax(probs)+1 
    
    # update the confusion matrix for reviews that are not having score as 3.
    if score != 3:
        update_confusion_matrix(pred_score, score)
    

## Confusion Matrix

confusion_matrix

recall = confusion_matrix['TP']/(confusion_matrix['TP'] + confusion_matrix['FN'])
recall

##  PRECISION

$ precision  = \frac{TP} {TP + FP} $

precision = confusion_matrix['TP']/(confusion_matrix['TP'] + confusion_matrix['FP'])
print('Precision: {}'.format(precision))

## RECALL 

$ recall = \frac{TP}{TP + FN}$

recall = confusion_matrix['TP']/(confusion_matrix['TP']+confusion_matrix['FN'])
print('recall: {}'.format(recall))

## Accuracy of our model (F score as metric)

+ We use $F_1$ score or simply F-score as our accuracy with testing data
+ basically $F_1$ score is _harmonic mean_ of _precision_ and _recall_

$F_1 = 2 . \frac{precision . recall}{precision + recall}$

f_score = 2*(precision*recall)/(precision+recall)
print('F score : {}'.format(f_score))
