==> 07/PART 3 NB from SKLearn.ipynb <==
import pandas as pd
import numpy as np

reviews_sst = pd.read_csv('preprocessed_reviews_SST.csv')

reviews_sst[:4]

dataset = pd.DataFrame()

dataset['smry_txt'] = reviews_sst['Summary'].astype(str) + ' ' + reviews_sst['Text']

dataset['score'] = reviews_sst['Score']

dataset.head()

dataset.shape

#lets filter the dataset so that it contains reviews that are either positive(4 or 5) or negative(1 or 2)
dataset = dataset[dataset.score != 3]

print(dataset.shape)
print('{} rows has review as 3'.format(568454 - 525814))

dataset['score'] = dataset['score'].apply({1:'negative', 2:'negative', 4:'positive', 5:'positive'}.get)

dataset.head()

## split the training(70%) and testing(30%) data

import random
def getTrainAndTestIndices(size, trn_pcnt = 0.70, tst_pcnt = 0.30):
    print("\n we have {} datapoints".format(size), end = ". ")
    train_size = int(size*trn_pcnt)
    test_size = size - train_size
    print("Out of which, we will get {} points in our training data and {} points in our testing data. \n".format(train_size, test_size))
    ind = list(range(size))
    random.shuffle(ind)
    return ind[:train_size], ind[train_size:]

 trn_ind, tst_ind = getTrainAndTestIndices(dataset.shape[0])

len(trn_ind), len(tst_ind)

train_data = dataset.iloc[trn_ind]
test_data = dataset.iloc[tst_ind]
train_data.shape, test_data.shape 

train_data.head(2)

## creating word count Sparse vector representation of Training data 

from sklearn.feature_extraction.text import CountVectorizer
count_vectorizer = CountVectorizer()

word_counts = count_vectorizer.fit_transform(train_data['smry_txt'].values)

type(word_counts), word_counts

# lets see how sparse matrix looks like
# print(word_counts[0])

from sklearn.naive_bayes import MultinomialNB

x_test, y_test = test_data['smry_txt'].values, test_data['score'].values

multi_nb_classifier = MultinomialNB()
multi_nb_classifier.fit(X=word_counts, y=train_data['score'])

pred = multi_nb_classifier.predict(count_vectorizer.transform(x_test))

pred[:10], y_test[:10]

accuracy = sum(pred == np.array(y_test)) / len(pred)
print('accuracy : {}'.format(accuracy))

from sklearn.metrics import classification_report
print(classification_report(y_test, pred))

from sklearn.metrics import precision_recall_fscore_support
precision_recall_fscore_support(y_test, pred, average='micro')

## Calculating F1_score using function

from collections import Counter
# intialize confusion matrix
confusion_matrix = Counter()

def update_confusion_matrix(predicted, actual):
    
#     if(predicted == 'positive' and actual == 'positive'):
#         print(predicted, actual)
    if predicted == actual and actual == 'positive':
        # update true positives
        confusion_matrix.update({'TP':1})
    elif predicted == actual and actual == 'negative':
        # update true negatives
        confusion_matrix.update({'TN':1})
    elif predicted != actual and actual == 'positive':
        # update false negatives
        confusion_matrix.update({'FN':1})
    elif predicted != actual and actual == 'negative':
        # update false positives
        confusion_matrix.update({'FP':1})

confusion_matrix.clear()
for tp in zip(pred, y_test):
    update_confusion_matrix(str(tp[0]), str(tp[1]))
confusion_matrix

precision = confusion_matrix['TP']/(confusion_matrix['TP'] + confusion_matrix['FP'])
print('Precision:{0:.12f}'.format(precision))

recall = confusion_matrix['TP']/(confusion_matrix['TP']+confusion_matrix['FN'])
print('recall: {0:.12f}'.format(recall))

f_score = 2*(precision*recall)/(precision+recall)
print('F score : {0:.12f}'.format(f_score))
