==> 07/part-5 SVM_KAGGLE_FINAL.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

reviews = pd.read_csv('../input/Reviews.csv')
reviews.head(1)

reviews['smry_txt'] = reviews['Summary'].astype(str) + ' ' + reviews['Text']
dataset = reviews.loc[reviews['Score'] != 3][['smry_txt','Score']]

dataset.loc[dataset['Score']==1, 'Score'] = 'negative'
dataset.loc[dataset['Score']==2, 'Score'] = 'negative'
dataset.loc[dataset['Score']==4, 'Score'] = 'positive'
dataset.loc[dataset['Score']==5, 'Score'] = 'positive'
dataset.head()

from sklearn.feature_extraction.text import CountVectorizer
count_vec_total = CountVectorizer()

x_data = count_vec_total.fit_transform(dataset.smry_txt)
y_data = dataset['Score']

y_data.head()

print(dataset.iloc[0]['smry_txt'], end='\n\n')
print(count_vec_total.inverse_transform(x_data[0]), end='\n\n')
print(x_data[0])

# split the data into training and testing

from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(x_data, y_data)

y_train.shape

# Linear svm with Gridsearch

from sklearn.model_selection import GridSearchCV
from sklearn.svm import LinearSVC

param_grid = {'C':[0.03125, 0.125, 0.5, 2, 8]}
lr_svm = LinearSVC(verbose=20, max_iter = 100000)
lr_svm_grid_cv = GridSearchCV(lr_svm, param_grid=param_grid, n_jobs=-1, verbose=20)
lr_svm_grid_cv

from datetime import datetime
start_time = datetime.now()

#perform grid search for c
lr_svm_grid_cv.fit(x_train, y_train)

print(datetime.now() - start_time)

lr_svm_grid_cv.best_params_, lr_svm_grid_cv.best_estimator_, lr_svm_grid_cv.best_score_

y_true, y_pred = y_test, lr_svm_grid_cv.predict(x_test)

from sklearn.metrics import classification_report, precision_recall_fscore_support
print(classification_report(y_true, y_pred))
print(precision_recall_fscore_support(y_true, y_pred, average='micro'))

from sklearn.metrics import roc_curve, auc, roc_auc_score

y_pred = np.where(y_pred == 'positive', 1, 0)
y_true = np.where(y_true == 'positive', 1, 0)

fpr, tpr, threshold = roc_curve(y_true, y_pred)
print("FPR: {}, \nTPR: {}, \nAUC : {}".format(fpr[1], tpr[1], auc(fpr,tpr)))

# SVM with RBF kernel

from sklearn.svm import SVC

param_grid = {'C':[0.03125, 0.125, 0.5, 2, 8],
             'gamma': [3.0517578125e-05, 0.00048828125, 0.0078125, 0.125, 2]}
rbf_svm = SVC(kernel='rbf')
rbf_svm_grid_cv = GridSearchCV(rbf_svm, param_grid=param_grid, n_jobs=-1, verbose=30)
rbf_svm_grid_cv

from datetime import datetime
start_time = datetime.now()

#perform gridsearch
rbf_svm_grid_cv.fit(x_train, y_train)

print(datetime.now() - start_time)

rbf_svm_grid_cv.best_params_, rbf_svm_grid_cv.best_estimator_, rbf_svm_grid_cv.best_score_

y_true, y_pred = y_test, rbf_svm_grid_cv.predict(x_test)

from sklearn.metrics import precision_recall_fscore_support, classification_report
print(classification_report(y_true, y_pred))
print(precision_recall_fscore_support(y_true, y_pred, average='micro'))

from sklearn.metrics import roc_curve, auc, roc_auc_score

y_pred = np.where(y_pred == 'positive', 1, 0)
y_true = np.where(y_true == 'positive', 1, 0)

fpr, tpr, threshold = roc_curve(y_true, y_pred)
print("FPR: {}, \nTPR: {}, \nAUC : {}".format(fpr[1], tpr[1], auc(fpr,tpr)))
